+++ src/mrpro/utils/typing.py
@@ -1,8 +1,9 @@
 """Some type hints that are used in multiple places in the codebase but not part of mrpro's public API."""
 
+from collections.abc import Callable
 from typing import TYPE_CHECKING, TypeAlias
 
-from typing_extensions import Any, Protocol, TypeVar
+from typing_extensions import Any, ParamSpec, Protocol, TypeVar, Unpack, overload
 
 if TYPE_CHECKING:
     from types import EllipsisType
@@ -22,6 +23,192 @@ if TYPE_CHECKING:
     _SingleNumpyIndexerType: TypeAlias = ndarray | SupportsIndex | None | slice | EllipsisType
     NumpyIndexerType: TypeAlias = tuple[_SingleNumpyIndexerType, ...] | _SingleNumpyIndexerType
 
+    Tout = TypeVar('Tout', bound=tuple[torch.Tensor, ...], covariant=True)
+
+    P = ParamSpec('P')
+    Wrapped: TypeAlias = Callable[P, Tout]
+
+    F = TypeVar('F', bound=Wrapped)
+
+    class _EndomorphCallable(Protocol):
+        """A callable with the same number of tensor inputs and outputs.
+
+        This is a protocol for a callable that takes a variadic number of tensor inputs
+        and returns the same number of tensor outputs.
+
+        This is only implemented for up to 10 inputs, if more inputs are given, the return
+        will be a variadic number of tensors.
+
+        This Protocol is used to decorate functions with the `endomorph` decorator.
+        """
+
+        @overload
+        def __call__(self, /) -> tuple[()]: ...
+        @overload
+        def __call__(self, x1: torch.Tensor, /) -> tuple[torch.Tensor]: ...
+
+        @overload
+        def __call__(self, x1: torch.Tensor, x2: torch.Tensor, /) -> tuple[torch.Tensor, torch.Tensor]: ...
+
+        @overload
+        def __call__(
+            self, x1: torch.Tensor, x2: torch.Tensor, x3: torch.Tensor, /
+        ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]: ...
+
+        @overload
+        def __call__(
+            self, x1: torch.Tensor, x2: torch.Tensor, x3: torch.Tensor, x4: torch.Tensor, /
+        ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]: ...
+
+        @overload
+        def __call__(
+            self, x1: torch.Tensor, x2: torch.Tensor, x3: torch.Tensor, x4: torch.Tensor, x5: torch.Tensor, /
+        ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]: ...
+
+        @overload
+        def __call__(
+            self,
+            x1: torch.Tensor,
+            x2: torch.Tensor,
+            x3: torch.Tensor,
+            x4: torch.Tensor,
+            x5: torch.Tensor,
+            x6: torch.Tensor,
+            /,
+        ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]: ...
+
+        @overload
+        def __call__(
+            self,
+            x1: torch.Tensor,
+            x2: torch.Tensor,
+            x3: torch.Tensor,
+            x4: torch.Tensor,
+            x5: torch.Tensor,
+            x6: torch.Tensor,
+            x7: torch.Tensor,
+            /,
+        ) -> tuple[
+            torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor
+        ]: ...
+
+        @overload
+        def __call__(
+            self,
+            x1: torch.Tensor,
+            x2: torch.Tensor,
+            x3: torch.Tensor,
+            x4: torch.Tensor,
+            x5: torch.Tensor,
+            x6: torch.Tensor,
+            x7: torch.Tensor,
+            x8: torch.Tensor,
+            /,
+        ) -> tuple[
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+        ]: ...
+
+        @overload
+        def __call__(
+            self,
+            x1: torch.Tensor,
+            x2: torch.Tensor,
+            x3: torch.Tensor,
+            x4: torch.Tensor,
+            x5: torch.Tensor,
+            x6: torch.Tensor,
+            x7: torch.Tensor,
+            x8: torch.Tensor,
+            x9: torch.Tensor,
+            /,
+        ) -> tuple[
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+        ]: ...
+
+        @overload
+        def __call__(
+            self,
+            x1: torch.Tensor,
+            x2: torch.Tensor,
+            x3: torch.Tensor,
+            x4: torch.Tensor,
+            x5: torch.Tensor,
+            x6: torch.Tensor,
+            x7: torch.Tensor,
+            x8: torch.Tensor,
+            x9: torch.Tensor,
+            x10: torch.Tensor,
+            /,
+        ) -> tuple[
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+        ]: ...
+
+        @overload
+        def __call__(
+            self,
+            x1: torch.Tensor,
+            x2: torch.Tensor,
+            x3: torch.Tensor,
+            x4: torch.Tensor,
+            x5: torch.Tensor,
+            x6: torch.Tensor,
+            x7: torch.Tensor,
+            x8: torch.Tensor,
+            x9: torch.Tensor,
+            x10: torch.Tensor,
+            /,
+            *args: torch.Tensor,
+        ) -> tuple[
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            torch.Tensor,
+            Unpack[tuple[torch.Tensor, ...]],
+        ]: ...
+
+        @overload
+        def __call__(self, /, *args: torch.Tensor) -> tuple[torch.Tensor, ...]: ...
+
+        def __call__(self, /, *args: torch.Tensor) -> tuple[torch.Tensor, ...]:
+            """Apply the Operator."""
+
+    def endomorph(f: F, /) -> _EndomorphCallable:
+        """Decorate a function to make it an endomorph callable.
+
+        This adds overloads for N->N-Tensor signatures, for N<10.
+        For >10 inputs, the return type will a tuple with >10 tensors.
+        """
+        return f
 
 else:
     TorchIndexerType: TypeAlias = Any
@@ -35,4 +222,9 @@ else:
     NumpyIndexerType: TypeAlias = Any
     """Numpy indexer type."""
 
-__all__ = ['NestedSequence', 'NumpyIndexerType', 'TorchIndexerType']
+    def endomorph(f: F) -> F:
+        """Decorate a function to make it an endomorph callable."""
+        return F
+
+
+__all__ = (['NestedSequence', 'NumpyIndexerType', 'TorchIndexerType', 'endomorph'],)
